move 'Radio' to 'src/components' folder
