+++ src/modules/pages/NewOAuthServerPage/OAuthServerForm.js
@@ -16,7 +16,7 @@ import Section from '../../Layout/Section/Section';
 import Row from '../../Layout/Row/Row';
 import Title from '../../Layout/Title/Title';
 import Input from '../../inputs/Input';
-import Radio from '../../inputs/Radio/Radio';
+import Radio from '../../../components/Radio/Radio';
 import Label from '../../labels/Label';
 import Hint from '../../../components/Hint/Hint';
 import SimpleSelect from '../../selects/SimpleSelect/SimpleSelect';
