+++ src/modules/pages/NewOAuthServerPage/partials/OAuthClientEndpoints/OAuthClientEndpoint.js
@@ -11,7 +11,7 @@ import optionsTransformer from '../../../../../helpers/optionsTransformer';
 
 import Row from '../../../../Layout/Row/Row';
 import Label from '../../../../labels/Label';
-import Radio from '../../../../inputs/Radio/Radio';
+import Radio from '../../../../../components/Radio/Radio';
 import Input from '../../../../inputs/Input';
 import Hint from '../../../../../components/Hint/Hint';
 import MultiSelect from '../../../../selects/MultiSelect/MultiSelect';
