+++ src/modules/pages/NewOAuthServerPage/partials/OAuthEndpoints/OAuthEndpoint.js
@@ -12,7 +12,7 @@ import getValues from '../../../../../helpers/getValues';
 
 import Row from '../../../../Layout/Row/Row';
 import Label from '../../../../labels/Label';
-import Radio from '../../../../inputs/Radio/Radio';
+import Radio from '../../../../../components/Radio/Radio';
 import Input from '../../../../inputs/Input';
 import Hint from '../../../../../components/Hint/Hint';
 import MultiSelect from '../../../../selects/MultiSelect/MultiSelect';
